+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.1
 
 - Initial connector to replace MongoDB connector in Tyk
+- Added support for uptime purging (redis/mongo only)